Add files via upload
tried to create an interface which will give you response based on the query from the user. it needs to be kept in mind that it will first navigate all the files given for training to give response and then it moves to the api's to fetch other data related to the query from the user. it also recommend YouTube video for the users query.